Compare methods moved to ComparisonUtilities
