+++ tests/Plivo.Test/Resources/TestCall.cs
@@ -7,6 +7,7 @@ using Plivo.Client;
 using Plivo.Http;
 using Plivo.Resource;
 using Plivo.Resource.Call;
+using Plivo.Utilities;
 
 namespace Plivo.Test.Resources
 {
@@ -38,7 +39,7 @@ namespace Plivo.Test.Resources
             Setup<CallCreateResponse>(201, response);
             
             Assert.IsEmpty(
-                Util.Compare(
+                ComparisonUtilities.Compare(
                     response, 
                     Api.Call.Create(
                         "+919999999999", 
@@ -98,7 +99,7 @@ namespace Plivo.Test.Resources
             Setup<ListResponse<Call>>(200, response);
             
             Assert.IsEmpty(
-                Util.Compare(
+                ComparisonUtilities.Compare(
                     response, 
                     Api.Call.List(endTime_Gt : DateTime.Parse("2017-06-07"))
                 )
@@ -126,7 +127,7 @@ namespace Plivo.Test.Resources
             Setup<Call>(200, response);
             
             Assert.IsEmpty(
-                Util.Compare(
+                ComparisonUtilities.Compare(
                     response, 
                     Api.Call.Get(id)
                 )
@@ -155,7 +156,7 @@ namespace Plivo.Test.Resources
             Setup<LiveCallListResponse>(200, response);
             
             Assert.IsEmpty(
-                Util.Compare(
+                ComparisonUtilities.Compare(
                     response, 
                     Api.Call.ListLive()
                 )
@@ -187,7 +188,7 @@ namespace Plivo.Test.Resources
             Setup<LiveCall>(200, response);
             
             Assert.IsEmpty(
-                Util.Compare(
+                ComparisonUtilities.Compare(
                     response, 
                     Api.Call.GetLive(id)
                 )
@@ -221,7 +222,7 @@ namespace Plivo.Test.Resources
             Setup<UpdateResponse<Call>>(202, response);
             
             Assert.IsEmpty(
-                Util.Compare(
+                ComparisonUtilities.Compare(
                     response, 
                     Api.Call.Transfer(id, "aleg", "http://asdsa.asdsa")
                 )
@@ -257,7 +258,7 @@ namespace Plivo.Test.Resources
             Setup<UpdateResponse<Call>>(202, response);
             
             Assert.IsEmpty(
-                Util.Compare(
+                ComparisonUtilities.Compare(
                     response, 
                     Api.Call.StartRecording(id, 120, "wav", callbackMethod: "http://a.a", callbackUrl: "http://s.s")
                 )
@@ -291,7 +292,7 @@ namespace Plivo.Test.Resources
             Setup<UpdateResponse<Call>>(202, response);
             
             Assert.IsEmpty(
-                Util.Compare(
+                ComparisonUtilities.Compare(
                     response, 
                     Api.Call.StartSpeaking(id, "120", voice: "WOMAN")
                 )
@@ -324,7 +325,7 @@ namespace Plivo.Test.Resources
             Setup<UpdateResponse<Call>>(202, response);
             
             Assert.IsEmpty(
-                Util.Compare(
+                ComparisonUtilities.Compare(
                     response, 
                     Api.Call.StartPlaying(id, new List<string>(){"http://wewewe.ewewew","http:/second.url"})
                 )
@@ -403,7 +404,7 @@ namespace Plivo.Test.Resources
             Console.WriteLine(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location));
             
             Assert.IsEmpty(
-                Util.Compare(
+                ComparisonUtilities.Compare(
                     response,
                     Api.Call.StopSpeaking(id)));
