+++ tests/Plivo.Test/Resources/TestMessage.cs
@@ -3,6 +3,7 @@ using NUnit.Framework;
 using Plivo.Http;
 using Plivo.Resource;
 using Plivo.Resource.Message;
+using Plivo.Utilities;
 
 namespace Plivo.Test.Resources
 {
@@ -35,7 +36,7 @@ namespace Plivo.Test.Resources
                 response
             );
             Assert.IsEmpty(
-                Util.Compare(
+                ComparisonUtilities.Compare(
                     response,
                     Api.Message.Create("+919999999999", new List<string>(){"+919898989898", "+919090909090"}, "textext")));
             AssertRequest(request);
@@ -64,7 +65,7 @@ namespace Plivo.Test.Resources
                 response
             );
             Assert.IsEmpty(
-                Util.Compare(
+                ComparisonUtilities.Compare(
                     response,
                     Api.Message.List(limit:10)));
             
@@ -90,7 +91,7 @@ namespace Plivo.Test.Resources
                 response
             );
             Assert.IsEmpty(
-                Util.Compare(
+                ComparisonUtilities.Compare(
                     response,
                     Api.Message.Get(id)));
